requirements.txt
pandas-1.4.3
